Fix to allow mbed header patching to work on Windows 7.
Windows 7 wouldn't allow patch.exe to run unless you were running it from a
process at elevated privilege.  This is because programs which contain the
name patch must be run from this state.  I simply renamed it to applydiff
instead and now Windows 7 is happy with it.

I also updated the README to indicate which installers should be used for
Code Sourcery setup.