Closes #4  Delayed stdio handle open causes issues.
Problem was reported by Charles Garcia-Tobin on mbed forum.

I have reverted my code to open stdin/stdout/stderr before global constructors
are run like I did originally.  This increases the size of some samples such
as HelloWorld which don't require these handles since the resulting call to
_sys_open() results in a call to sscanf() which pulls in another ~20k of code.
I have added a GCC4MBED_DELAYED_STDIO_INIT makefile variable which can be set
to a non-zero value for samples like HelloWorld to revert to the old behaviour
which will still do the delayed open.  This means that if a developer happens
to set the GCC4MBED_DELAYED_STDIO_INIT by mistake, they will still get some
output on UART0 but it might not be of the expected format.  This is better
than the behaviour you get if the files aren't even opened.