Closes #5 Force cs-make to use cmd.exe as its shell on Windows
Previously, cs-make would use /usr/bin/sh if it found Cygwin in the path.
The gcc4mbed makefiles were written to expect Windows cmd.exe to be used
as the shell on Windows so that the user wouldn't need Cygwin.

Added SHELL=cmd.exe to makefiles when cs-make is used as the make utility.