Wrap semihost_connected() API when MRI is enabled.
The mbed SDK assumes that there is no way to obtain the MAC address from the
interface chip when the JTAG interface is shutdown by MRI.  This isn't
completely true since MRI implements this semihost call even when the
interface chip has been disconnected.  When MRI is enabled, the call to
semihost_connected() is wrapped and forced to return -1, indicating that
the interface is still up.

This has one other side effect.  Previously calls like exit() would issue a
semihost breakpoint if the mbed interface was connected but just enter an
infinite loop when MRI had disabled it.  Now they issue the semihost breakpoint
which gets caught by MRI and since it has no special handling for these
particular semihost calls, it will just break into the debugger.  This is
actually a side effect that I like since it lets the developer see that
there program has called exit and their debug session is really over.