Fixed typo in build/gcc4mbed.mk
Noticed when code reviewing a merge.  The MRI_WRAPS make variable is used to
add linker wraps for routines that must be hooked when debugging.  When MRI
isn't enabled, it is supposed to be forced to be an empty string but it was
misspelled as MRI_WRAP in this scenario.