Fix recent assembly language changes to work on Windows.
On *nix, the wildcard function used to populate ASRCS is case sensitive but it
isn't case sensitive on Windows.  This meant that .s and .S files were pulled
in at once and then the patsubst used to convert the paths for the OBJECTS
variable would leave some of the original filenames around.

My fix was to always pull .s and .S filenames into ASRCS.  On Windows, we only
need to just use the wildcard with *.S but on *nix, the *.s list is added
explicitly.  The two patsubst calls were made to convert ASRCS from *.s/*.S to
*.o filenames for the OBJECTS variable.