Initial attempt to update windows and linux installers.
These updates to the newer GCC install still need to be tested.  This
commit just contains a port of the changes that I made for OS X.  I
also removed the .md5 files and not create them on the fly from the
install scripts.  It appears that the windows installers can just
extract binaries from a .zip archive and no longer needs a separate
archive just for the Windows binaries as I did previously.