Refactor makefiles to reduce redundancy
Pulled repeated code such as that for setting the compiler flags out of
many makefiles and placed once in device-libs.mk

Now just assemble all assembly language source files via GCC so that
they can be pre-processed if need be.  Now just support the AS_FLAGS
variable in the user's makefile and dropped AS_GCFLAGS.

A lot of the code in the makefiles which build libraries (mbed.a,
rtos.a, etc) had a lot of code in common.  I have pulled this out into
lib-common.mk and include from other library creation makefiles.