Allow switch between newlib-nano and newlib
The makefiles still defaults to using newlib-nano to produce smaller
binaries but the user can set the NEWLIB_NANO to 0 and the standard
newlib C library will be used instead.  One of the reasons I have added
this support is for people who want 64-bit integer support from the
printf family of functions which newlib-nano does not have.