Reduce supported device knowledge
Previously the build/gcc4mbed.mk and samples/makefile makefiles each
contained an explicit list of the supported devices.  This means that
when a new device was added, these makefiles would have to be updated
as well.

I renamed the device specific makefiles to have a -device.mk suffix so
that the wildcard function could be used to discover all supported
devices.  This means that the developer only needs to add one of these
-device.mk files to the build/ directory and the other makefiles will
automatically know about the newly added device.