Fix build to work on Windows
I had switched the recurse_dir makefile function to use the Posix find
command.  This works on OS X and Linux but not Windows.  On Windows, I
created a win32_find makefile function to substitute for find.  It uses
"dir /s" at its core but also needs to handle the fact that the
directory names in the results are supposed to be relative and not
absolute.