Update Cortex-M4F -device.mk files
The C_FLAGS, ASM_FLAGS, and LD_FLAGS all need to contain the following
CPU/FPU options to fully select the type of device the code is being
built for:
    -mcpu=cortex-m4 -mthumb -mfpu=fpv4-sp-d16 -mfloat-abi=softfp
For the compiler, these flags indicate the type of code it should
generate.  For the linker, the flags indicate which standard C/C++
library version should be pulled from the multi-lib.