Updating to GCC ARM Embedded 4.9-2014-q4-major
I have updated and tested the mac_install script to install and use the
latest build of GCC ARM Embedded.  I also updated the Linux and Windows
install scripts too but I will test them later when I am closer to
being ready for merging into master.

I also updated the README to note the newer version of GCC and added
a newer version of GDB for OS X which includes my fixes to work with
the mbed-LPC1768 CDC firmware.