Removing link to oldest version
The README used to have links to all older versions of itself,
including the first one which used Code Sourcery before I switched to
GNU Tools for ARM Embedded. I want to get rid of the supporting
GitHub repository that I have for the Code Sourcery OS X binaries and
the removal of this link is the first step.