Add Freescale K64F support
This commit adds support for the FRDM-K64F board. I also updated the
makefiles for HelloWorld, StdIO, Ticker, rtos_basic, and
TCPSocket_HelloWorld samples to build for this platform as well.

I did try to enable the USBMouse sample as well but it fails to
enumerate. I need to debug that issue more before enabling it in the
build.