Add mbedUpdater to create *-device.mk files
The mbedUpdater application uses /external/mbed-os/targets/targets.json
to generate the required TARGET-device.mk files so that they no longer
need to be generated by hand.

It also produces a few TestPass*.mk files for the HelloWorld sample to
set DEVICES to all of the targets that mbed officially supports for
version 2 and 5 of the library. I can use these makefiles to make sure
that all officially supported targets at least build with GCC4MBED.