Allow USER_LIBS to not recurse for include path
If you now prefix a library folder in the USER_LIBS list with a '!'
character then just the root of the library folder will be added to
include path instead of every subfolder within the library project.

I also added a note to the MBED_CONFIG_H documentation to indicate the
risks of its use.